requirements.txt
streamlit
requests
pandas
newsapi-python
st-gsheets-connection